Unify fill functions into fill.py
